Move QualifiedSchemaName to SPI as CatalogSchemaName
This class is needed for schema operations in SystemAccessControl.
