+++ presto-main/src/main/java/com/facebook/presto/metadata/Metadata.java
@@ -15,6 +15,7 @@ package com.facebook.presto.metadata;
 
 import com.facebook.presto.Session;
 import com.facebook.presto.connector.ConnectorId;
+import com.facebook.presto.spi.CatalogSchemaName;
 import com.facebook.presto.spi.ColumnHandle;
 import com.facebook.presto.spi.ColumnMetadata;
 import com.facebook.presto.spi.ConnectorTableMetadata;
@@ -50,7 +51,7 @@ public interface Metadata
 
     void addFunctions(List<? extends SqlFunction> functions);
 
-    boolean schemaExists(Session session, QualifiedSchemaName schema);
+    boolean schemaExists(Session session, CatalogSchemaName schema);
 
     @NotNull
     List<String> listSchemaNames(Session session, String catalogName);
